Added the Player movement, updating the Room asset set to function with it, added the base weapon class.

+++ Assets/Scripts/Entities/ITurnActor.cs
@@ -3,7 +3,7 @@ public interface ITurnActor
 {
     void TakeTurn();
 
-    int GetSpeed();
+    float GetSpeed();
 
     bool IsAlive();
 }
